+++ src/custom/ecospheres/components/BouquetList.vue
@@ -39,13 +39,13 @@
 import { onMounted } from 'vue'
 import { useLoading } from 'vue-loading-overlay'
 
-import type { Topic } from '../model'
-import { NoOptionSelected } from '../model'
-import { useTopicStore } from '../store/TopicStore'
-import Tile from './Tile.vue'
+import Tile from '@/components/Tile.vue'
+import type { Topic } from '@/model'
+import { NoOptionSelected } from '@/model'
+import { useTopicStore } from '@/store/TopicStore'
 
 export default {
-  name: 'TopicList',
+  name: 'BouquetList',
   components: {
     Tile: Tile
   },
