+++ src/services/routerUtils.js
@@ -1,12 +1,12 @@
 import config from '@/config'
+import BouquetDetailView from '@/custom/ecospheres/views/bouquets/BouquetDetailView.vue'
+import BouquetEditView from '@/custom/ecospheres/views/bouquets/BouquetEditView.vue'
+import BouquetsListView from '@/custom/ecospheres/views/bouquets/BouquetsListView.vue'
 
 import HomeView from '../views/HomeView.vue'
 import LoginView from '../views/LoginView.vue'
 import LogoutView from '../views/LogoutView.vue'
 import SimplePageView from '../views/SimplePageView.vue'
-import BouquetDetailView from '../views/bouquets/BouquetDetailView.vue'
-import BouquetEditView from '../views/bouquets/BouquetEditView.vue'
-import BouquetsListView from '../views/bouquets/BouquetsListView.vue'
 import DatasetDetailView from '../views/datasets/DatasetDetailView.vue'
 import DatasetsListView from '../views/datasets/DatasetsListView.vue'
 import OrganizationDetailView from '../views/organizations/OrganizationDetailView.vue'
